backend/requirements.txt
# Sentinel-RX Backend Dependencies
# ================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
geoalchemy2==0.14.2

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Validation & Serialization
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML Libraries
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.7.2
pandas==2.1.3
numpy==1.26.2
networkx==3.2.1
ultralytics==8.0.200

# Voice Assistant
openai-whisper==20231117
gtts==2.4.0
openai==1.3.7

# Web Scraping & APIs
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
selenium==4.15.2

# Geolocation
geopy==2.4.1

# Translation
deep-translator==1.11.4

# Background Tasks
celery==5.3.4

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
python-dotenv==1.0.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
